+++ noodles-bcf/src/io/reader/record.rs
@@ -104,6 +104,7 @@ pub(crate) mod tests {
 
     #[test]
     fn test_read_record() -> Result<(), Box<dyn std::error::Error>> {
+        use noodles_core::Position;
         use noodles_vcf::variant::record_buf::{
             info::{self, field::Value as InfoFieldValue},
             samples::{
@@ -111,7 +112,7 @@ pub(crate) mod tests {
                 sample::{value::Array, Value as GenotypeFieldValue},
                 Keys,
             },
-            Position, Samples as VcfGenotypes,
+            Samples as VcfGenotypes,
         };
 
         use crate::header::StringMaps;
@@ -124,7 +125,10 @@ pub(crate) mod tests {
         read_record(&mut reader, &mut record)?;
 
         assert_eq!(record.chromosome_id()?, 1);
-        assert_eq!(record.position()?, Position::from(101));
+        assert_eq!(
+            record.position().transpose()?,
+            Some(Position::try_from(101)?)
+        );
         assert_eq!(record.rlen()?, 1);
         assert_eq!(record.quality_score()?, Some(30.1));
         assert_eq!(record.ids().as_ref(), b"rs123");