+++ noodles-vcf/examples/vcf_write.rs
@@ -4,10 +4,10 @@
 
 use std::io;
 
+use noodles_core::Position;
 use noodles_vcf::{
     self as vcf,
     header::record::value::{map::Contig, Map},
-    variant::record_buf::Position,
 };
 
 fn main() -> io::Result<()> {
@@ -22,7 +22,7 @@ fn main() -> io::Result<()> {
 
     let record = vcf::variant::RecordBuf::builder()
         .set_chromosome("sq0")
-        .set_position(Position::from(1))
+        .set_position(Position::MIN)
         .set_reference_bases("A")
         .build();